move tests for test deck utils from VxDesign to libs/utils
