+++ apps/design/backend/src/test_decks.test.ts
@@ -19,9 +19,8 @@ import {
   vxGeneralElectionFixtures,
   vxPrimaryElectionFixtures,
 } from '@votingworks/hmpb';
-import { assert, find, iter, ok } from '@votingworks/basics';
+import { assert, iter, ok } from '@votingworks/basics';
 import {
-  buildContestResultsFixture,
   CachedElectionLookups,
   generateTestDeckBallots,
   getBallotStyleGroupsForPrecinctOrSplit,
@@ -48,8 +47,6 @@ import {
   createPrecinctTestDeck,
   createPrecinctSummaryBallotTestDeck,
   createTestDeckTallyReports,
-  generateTestDeckCastVoteRecords,
-  getTallyReportResults,
   precinctTallyReportFileName,
   FULL_TEST_DECK_TALLY_REPORT_FILE_NAME,
 } from './test_decks';
@@ -345,309 +342,6 @@ describe('createTestDeckTallyReports', () => {
     expect(
       reports.has(precinctTallyReportFileName(singlePrecinct.name))
     ).toEqual(false);
-  });
-});
-
-describe('getTallyReportResults', () => {
-  test('general election without summary ballots', async () => {
-    const { electionDefinition } = vxFamousNamesFixtures;
-    const { election } = electionDefinition;
-
-    const cvrs = generateTestDeckCastVoteRecords(election, {
-      includeSummaryBallots: false,
-    });
-    const tallyReportResults = await getTallyReportResults(election, cvrs);
-
-    expect(tallyReportResults.hasPartySplits).toEqual(false);
-    expect(tallyReportResults.contestIds).toEqual(
-      election.contests.map((c) => c.id)
-    );
-    expect(tallyReportResults.manualResults).toBeUndefined();
-    const { scannedResults } = tallyReportResults;
-    // Without summary ballots, only HMPB counts
-    expect(scannedResults.cardCounts).toEqual({
-      bmd: [],
-      hmpb: [52],
-    });
-
-    // check one contest
-    expect(scannedResults.contestResults['board-of-alderman']).toEqual(
-      buildContestResultsFixture({
-        contest: find(election.contests, (c) => c.id === 'board-of-alderman'),
-        contestResultsSummary: {
-          type: 'candidate',
-          ballots: 52,
-          overvotes: 0,
-          undervotes: 156,
-          officialOptionTallies: {
-            'helen-keller': 8,
-            'nikola-tesla': 8,
-            'pablo-picasso': 4,
-            'steve-jobs': 8,
-            'vincent-van-gogh': 4,
-            'wolfgang-amadeus-mozart': 4,
-            'write-in': 16,
-          },
-        },
-        includeGenericWriteIn: true,
-      })
-    );
-  });
-
-  test('general election with summary ballots doubles the counts', async () => {
-    const { electionDefinition } = vxFamousNamesFixtures;
-    const { election } = electionDefinition;
-
-    const cvrs = generateTestDeckCastVoteRecords(election, {
-      includeSummaryBallots: true,
-    });
-    const tallyReportResults = await getTallyReportResults(election, cvrs);
-
-    expect(tallyReportResults.hasPartySplits).toEqual(false);
-    const { scannedResults } = tallyReportResults;
-    // With summary ballots, we have both HMPB and BMD counts (doubled)
-    expect(scannedResults.cardCounts).toEqual({
-      bmd: [52],
-      hmpb: [52],
-    });
-
-    // check one contest - tallies should be doubled
-    expect(scannedResults.contestResults['board-of-alderman']).toEqual(
-      buildContestResultsFixture({
-        contest: find(election.contests, (c) => c.id === 'board-of-alderman'),
-        contestResultsSummary: {
-          type: 'candidate',
-          ballots: 104, // doubled
-          overvotes: 0,
-          undervotes: 312, // doubled
-          officialOptionTallies: {
-            'helen-keller': 16, // doubled
-            'nikola-tesla': 16,
-            'pablo-picasso': 8,
-            'steve-jobs': 16,
-            'vincent-van-gogh': 8,
-            'wolfgang-amadeus-mozart': 8,
-            'write-in': 32,
-          },
-        },
-        includeGenericWriteIn: true,
-      })
-    );
-  });
-
-  test('primary election without summary ballots', async () => {
-    const { electionDefinition } = vxPrimaryElectionFixtures;
-    const { election } = electionDefinition;
-
-    const cvrs = generateTestDeckCastVoteRecords(election, {
-      includeSummaryBallots: false,
-    });
-    const tallyReportResults = await getTallyReportResults(election, cvrs);
-
-    expect(tallyReportResults.hasPartySplits).toEqual(true);
-    expect(tallyReportResults.contestIds).toEqual(
-      election.contests.map((c) => c.id)
-    );
-    expect(tallyReportResults.manualResults).toBeUndefined();
-    expect(
-      tallyReportResults.hasPartySplits && tallyReportResults.cardCountsByParty
-    ).toEqual({
-      '0': {
-        bmd: [],
-        hmpb: [100],
-      },
-      '1': {
-        bmd: [],
-        hmpb: [100],
-      },
-    });
-    const { scannedResults } = tallyReportResults;
-    expect(scannedResults.cardCounts).toEqual({
-      bmd: [],
-      hmpb: [200],
-      manual: 0,
-    });
-
-    // check one contest
-    expect(scannedResults.contestResults['county-leader-mammal']).toEqual(
-      buildContestResultsFixture({
-        contest: find(
-          election.contests,
-          (c) => c.id === 'county-leader-mammal'
-        ),
-        contestResultsSummary: {
-          type: 'candidate',
-          ballots: 100,
-          overvotes: 0,
-          undervotes: 0,
-          officialOptionTallies: {
-            fox: 20,
-            horse: 40,
-            otter: 40,
-          },
-        },
-        includeGenericWriteIn: false,
-      })
-    );
-  });
-
-  test('primary election with summary ballots doubles the counts', async () => {
-    const { electionDefinition } = vxPrimaryElectionFixtures;
-    const { election } = electionDefinition;
-
-    const cvrs = generateTestDeckCastVoteRecords(election, {
-      includeSummaryBallots: true,
-    });
-    const tallyReportResults = await getTallyReportResults(election, cvrs);
-
-    expect(tallyReportResults.hasPartySplits).toEqual(true);
-    expect(
-      tallyReportResults.hasPartySplits && tallyReportResults.cardCountsByParty
-    ).toEqual({
-      '0': {
-        bmd: [100],
-        hmpb: [100],
-      },
-      '1': {
-        bmd: [100],
-        hmpb: [100],
-      },
-    });
-    const { scannedResults } = tallyReportResults;
-    expect(scannedResults.cardCounts).toEqual({
-      bmd: [200],
-      hmpb: [200],
-      manual: 0,
-    });
-
-    // check one contest - tallies should be doubled
-    expect(scannedResults.contestResults['county-leader-mammal']).toEqual(
-      buildContestResultsFixture({
-        contest: find(
-          election.contests,
-          (c) => c.id === 'county-leader-mammal'
-        ),
-        contestResultsSummary: {
-          type: 'candidate',
-          ballots: 200, // doubled
-          overvotes: 0,
-          undervotes: 0,
-          officialOptionTallies: {
-            fox: 40, // doubled
-            horse: 80,
-            otter: 80,
-          },
-        },
-        includeGenericWriteIn: false,
-      })
-    );
-  });
-
-  test('general election precinct-specific results', async () => {
-    const { electionDefinition } = vxFamousNamesFixtures;
-    const { election } = electionDefinition;
-    const precinct = election.precincts[0];
-
-    const cvrs = generateTestDeckCastVoteRecords(election, {
-      includeSummaryBallots: false,
-    });
-    const precinctCvrs = cvrs.filter((cvr) => cvr.precinctId === precinct.id);
-    const tallyReportResults = await getTallyReportResults(
-      election,
-      precinctCvrs,
-      precinct.id
-    );
-
-    expect(tallyReportResults.hasPartySplits).toEqual(false);
-    // Precinct-specific results only include contests for that precinct
-    expect(tallyReportResults.contestIds.length).toEqual(8);
-    const { scannedResults } = tallyReportResults;
-    expect(scannedResults.cardCounts).toEqual({
-      bmd: [],
-      hmpb: [13],
-    });
-
-    // check one contest - should have precinct-specific counts
-    expect(scannedResults.contestResults['board-of-alderman']).toEqual(
-      buildContestResultsFixture({
-        contest: find(election.contests, (c) => c.id === 'board-of-alderman'),
-        contestResultsSummary: {
-          type: 'candidate',
-          ballots: 13,
-          overvotes: 0,
-          undervotes: 39,
-          officialOptionTallies: {
-            'helen-keller': 2,
-            'nikola-tesla': 2,
-            'pablo-picasso': 1,
-            'steve-jobs': 2,
-            'vincent-van-gogh': 1,
-            'wolfgang-amadeus-mozart': 1,
-            'write-in': 4,
-          },
-        },
-        includeGenericWriteIn: true,
-      })
-    );
-  });
-
-  test('primary election precinct-specific results', async () => {
-    const { electionDefinition } = vxPrimaryElectionFixtures;
-    const { election } = electionDefinition;
-    const precinct = election.precincts[0];
-
-    const cvrs = generateTestDeckCastVoteRecords(election, {
-      includeSummaryBallots: false,
-    });
-    const precinctCvrs = cvrs.filter((cvr) => cvr.precinctId === precinct.id);
-    const tallyReportResults = await getTallyReportResults(
-      election,
-      precinctCvrs,
-      precinct.id
-    );
-
-    expect(tallyReportResults.hasPartySplits).toEqual(true);
-    assert(tallyReportResults.hasPartySplits);
-    // Precinct-specific results only include contests for that precinct
-    expect(tallyReportResults.contestIds.length).toEqual(5);
-    expect(tallyReportResults.cardCountsByParty).toEqual({
-      '0': {
-        bmd: [],
-        hmpb: [20],
-      },
-      '1': {
-        bmd: [],
-        hmpb: [20],
-      },
-    });
-    const { scannedResults } = tallyReportResults;
-    expect(scannedResults.cardCounts).toEqual({
-      bmd: [],
-      hmpb: [40],
-      manual: 0,
-    });
-
-    // check one contest - should have precinct-specific counts
-    expect(scannedResults.contestResults['county-leader-mammal']).toEqual(
-      buildContestResultsFixture({
-        contest: find(
-          election.contests,
-          (c) => c.id === 'county-leader-mammal'
-        ),
-        contestResultsSummary: {
-          type: 'candidate',
-          ballots: 20,
-          overvotes: 0,
-          undervotes: 0,
-          officialOptionTallies: {
-            fox: 4,
-            horse: 8,
-            otter: 8,
-          },
-        },
-        includeGenericWriteIn: false,
-      })
-    );
   });
 });
